requirements.txt
streamlit
pillow
xhtml2pdf
